[WIP] Merge TVLockO code with the Quiz code to merge their functionalities
