+++ app/src/main/AndroidManifest.xml
@@ -30,28 +30,16 @@
             android:screenOrientation="landscape" />
         <activity
             android:exported="true"
-            android:name=".HomeActivity"
-            android:screenOrientation="landscape"
-            android:theme="@style/AppTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LEANBACK_LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:exported="true"
-            android:name="jp.funx.tekapo.TVLockActivity"
+            android:name=".MainActivity"
             android:banner="@drawable/myapp_icon_foreground"
             android:icon="@drawable/myapp_icon_foreground"
             android:label="@string/app_name"
             android:logo="@drawable/myapp_icon_foreground"
             android:screenOrientation="landscape">
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LEANBACK_LAUNCHER" />-->
-<!--            </intent-filter>-->
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LEANBACK_LAUNCHER" />
+            </intent-filter>
         </activity>
         <activity
             android:exported="false"
@@ -60,7 +48,7 @@
             android:screenOrientation="landscape">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="jp.funx.tekapo.HomeActivity"/>
+                android:value="jp.funx.tekapo.MainActivity"/>
     </activity>
     <receiver
         android:name="jp.funx.tekapo.Receiver"
