+++ app/src/main/java/jp/funx/tekapo/Receiver.java
@@ -13,7 +13,7 @@ public class Receiver extends android.content.BroadcastReceiver {
         Log.d(TAG,"onReceive() - " + intent.getAction() );
 
         /*
-        Intent activityIntent = new Intent(context, TVLockActivity.class);
+        Intent activityIntent = new Intent(context, PINActivity.class);
         activityIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         context.startActivity(activityIntent);
         */
