+++ app/src/main/java/jp/funx/tekapo/SolutionActivity.java
@@ -52,7 +52,7 @@ public class SolutionActivity extends AppCompatActivity {
 
     @Override
     public void onBackPressed() {
-        Intent intent = new Intent(this,HomeActivity.class);
+        Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
     }
