+++ app/src/main/res/layout/activity_home.xml
@@ -7,7 +7,7 @@
     android:background="@color/colorPrimaryDark"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".HomeActivity">
+    tools:context=".MainActivity">
 
 
     <ImageView
